Created manager to handle connection status messages
