+++ Assets/BossRoom/Scenes/Startup.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:7f94d9ef30f99dd13a417106572a3d3dda71b434c4f9a41b7b54a717971bb6a8
-size 33791
+oid sha256:7b2c33d4681d53ec62b32b7ff2c81123b4457d8dcd8d0b6d643097e74ae36856
+size 35006
