+++ Assets/BossRoom/Scripts/Client/UI/ConnectionStatusMessageUIManager.cs
@@ -1,0 +1,53 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using Unity.Multiplayer.Samples.BossRoom;
+using Unity.Multiplayer.Samples.BossRoom.Shared.Infrastructure;
+using Unity.Multiplayer.Samples.BossRoom.Visual;
+using UnityEngine;
+
+public class ConnectionStatusMessageUIManager : MonoBehaviour
+{
+
+    IDisposable m_Subscriptions;
+
+    [Inject]
+    void InjectDependencies(ISubscriber<ConnectStatus> connectStatusSub)
+    {
+        m_Subscriptions = connectStatusSub.Subscribe(OnConnectStatus);
+    }
+
+    void Awake()
+    {
+        DontDestroyOnLoad(gameObject);
+    }
+
+    void OnDestroy()
+    {
+        m_Subscriptions?.Dispose();
+    }
+
+    void OnConnectStatus(ConnectStatus status)
+    {
+        switch (status)
+        {
+            case ConnectStatus.Undefined:
+            case ConnectStatus.UserRequestedDisconnect:
+                break;
+            case ConnectStatus.ServerFull:
+                PopupPanel.ShowPopupPanel("Connection Failed", "The Host is full and cannot accept any additional connections.");
+                break;
+            case ConnectStatus.Success:
+                break;
+            case ConnectStatus.LoggedInAgain:
+                PopupPanel.ShowPopupPanel("Connection Failed", "You have logged in elsewhere using the same account.");
+                break;
+            case ConnectStatus.GenericDisconnect:
+                PopupPanel.ShowPopupPanel("Disconnected From Host", "The connection to the host was lost");
+                break;
+            default:
+                Debug.LogWarning($"New ConnectStatus {status} has been added, but no connect message defined for it.");
+                break;
+        }
+    }
+}
