+++ Assets/BossRoom/Scripts/Client/UI/ConnectionStatusMessageUIManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6bf97fb8b52d37a4c99c1241500aea48
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
